python-backend/requirements.txt
# Core web framework
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
cryptography>=43.0.0

# Document processing
pypdf2>=3.0.0
python-docx>=1.1.0
pillow>=11.0.0
pytesseract>=0.3.13

# LLM and AI services
openai>=1.50.0
litellm>=1.50.0
langchain>=0.3.0
langchain-community>=0.3.0
llama-index>=0.12.0

# Vector store (may require pandas as dependency)
chromadb>=0.5.0

# Note: If pandas installation fails, install it separately first:
# pip install pandas --only-binary :all:
